+++ test/Feature/CleanRegex/PatternList/replace/withReferences/PatternListTest.php
@@ -1,12 +1,12 @@
 <?php
-namespace Test\Feature\CleanRegex\PatternList\chainedReplace\with;
+namespace Test\Feature\CleanRegex\PatternList\replace\withReferences;
 
 use PHPUnit\Framework\TestCase;
 use TRegx\CleanRegex\Pattern;
 use TRegx\Exception\MalformedPatternException;
 
 /**
- * @covers \TRegx\CleanRegex\PatternList::chainedReplace
+ * @covers \TRegx\CleanRegex\PatternList::replace
  * @covers \TRegx\CleanRegex\Composite\ChainedReplace
  */
 class PatternListTest extends TestCase
@@ -28,7 +28,7 @@ class PatternListTest extends TestCase
             '(\s+|\?)',
         ], 0, $times));
         // when
-        $replaced = $pattern->chainedReplace("Do you think that's air you're breathing now?")->with('__');
+        $replaced = $pattern->replace("Do you think that's air you're breathing now?")->withReferences('__');
         // then
         $this->assertSame($expected, $replaced);
     }
@@ -50,67 +50,67 @@ class PatternListTest extends TestCase
     /**
      * @test
      */
-    public function shouldReplacePcreReference()
+    public function shouldReplaceWithReferences()
     {
         // given
         $pattern = Pattern::list(['One(1)', 'Two(2)', 'Three(3)']);
         // when
-        $replaced = $pattern->chainedReplace('One1, Two2, Three3')->with('$1');
+        $replaced = $pattern->replace('One1 Two2 Three3')->withReferences('$1');
         // then
-        $this->assertSame('$1, $1, $1', $replaced);
+        $this->assertSame('1 2 3', $replaced);
     }
 
     /**
      * @test
      */
-    public function shouldReplaceWithPcreReferences_whole()
+    public function shouldReplaceWithReferences_whole()
     {
         // given
         $pattern = Pattern::list(['One(1)', 'Two(2)', 'Three(3)']);
         // when
-        $replaced = $pattern->chainedReplace('One1, Two2, Three3')->with('<$0>');
+        $replaced = $pattern->replace('One1, Two2, Three3')->withReferences('<$0>');
         // then
-        $this->assertSame('<$0>, <$0>, <$0>', $replaced);
+        $this->assertSame('<One1>, <Two2>, <Three3>', $replaced);
     }
 
     /**
      * @test
      */
-    public function shouldReplaceWithPcreReferences_whole_curlyBrace()
+    public function shouldReplaceWithReferences_whole_curlyBrace()
     {
         // given
         $pattern = Pattern::list(['One(1)', 'Two(2)', 'Three(3)']);
         // when
-        $replaced = $pattern->chainedReplace('One1, Two2, Three3')->with('<${0}>');
+        $replaced = $pattern->replace('One1, Two2, Three3')->withReferences('<${0}>');
         // then
-        $this->assertSame('<${0}>, <${0}>, <${0}>', $replaced);
+        $this->assertSame('<One1>, <Two2>, <Three3>', $replaced);
     }
 
     /**
      * @test
      */
-    public function shouldReplaceWithPcreReferences_whole_backslash()
+    public function shouldReplaceWithReferences_whole_backslash()
     {
         // given
         $pattern = Pattern::list(['One(1)', 'Two(2)', 'Three(3)']);
         // when
-        $replaced = $pattern->chainedReplace('One1, Two2, Three3')->with('<\0>');
+        $replaced = $pattern->replace('One1, Two2, Three3')->withReferences('<\0>');
         // then
-        $this->assertSame('<\0>, <\0>, <\0>', $replaced);
+        $this->assertSame('<One1>, <Two2>, <Three3>', $replaced);
     }
 
     /**
      * @test
      */
-    public function shouldReplaceWithPcreReferences_twoDigits_backslash()
+    public function shouldReplaceWithReferences_twoDigits_backslash()
     {
         // given
         $tenGroups = \str_repeat('()', 10);
         $pattern = Pattern::list(["One$tenGroups(1)", "Two$tenGroups(2)"]);
         // when
-        $replaced = $pattern->chainedReplace('One1, Two2')->with('<${11}>');
+        $replaced = $pattern->replace('One1, Two2')->withReferences('<${11}>');
         // then
-        $this->assertSame('<${11}>, <${11}>', $replaced);
+        $this->assertSame('<1>, <2>', $replaced);
     }
 
     /**
@@ -120,11 +120,11 @@ class PatternListTest extends TestCase
     {
         // given
         $patternList = Pattern::list(['\\']);
-        $replace = $patternList->chainedReplace('subject');
+        $replace = $patternList->replace('subject');
         // when
         $this->expectException(MalformedPatternException::class);
         $this->expectExceptionMessage('Pattern may not end with a trailing backslash');
         // when
-        $replace->with('replacement');
+        $replace->withReferences('replacement');
     }
 }